Scripts/Silverpine Forest: Fix Silverpine Forest till quest
There are several quests within the entire questline that need retouching yet due to a few inaccuracies. They will be deployed within the next days. DB fixes will be out soon for PR.

Please, feel free to advise or improve the current state of the zone in any possible way.