AOE Looting and Mailing Excess to Player.
This is a a attempt at a option core QOL (?) feature. Included is:
Conf to enable and to disabled the QOL.
AOE Looting (aka Mass Looting\MOP Looting) allows looting within a distance of 30 when NOT in a group.
Excess items are Mail to player when there is open spot in player inventory.
All items are looted. They are not Pick and chose when the conf is active.
This will be a limitation of the client \ core i assume.
Issues to not, When loot gets mailed to the player, it will do one message per a item. So at times the players inbox will be full due to client limitation and they will have to clear out some space and relog to refresh the remaining mail in the queue.